+++ t-fusionTool/src/test/java/eu/unifiedviews/plugins/transformer/fusiontool/testutils/MockEntry.java
@@ -2,20 +2,20 @@ package eu.unifiedviews.plugins.transformer.fusiontool.testutils;
 
 import eu.unifiedviews.dataunit.DataUnitException;
 import eu.unifiedviews.dataunit.rdf.RDFDataUnit;
-import org.openrdf.model.URI;
+import org.eclipse.rdf4j.model.IRI;
 
 /**
 *
 */
 public class MockEntry implements RDFDataUnit.Entry {
-    private final URI uri;
+    private final IRI uri;
 
-    public MockEntry(URI uri) {
+    public MockEntry(IRI uri) {
         this.uri = uri;
     }
 
     @Override
-    public URI getDataGraphURI() throws DataUnitException {
+    public IRI getDataGraphURI() throws DataUnitException {
         return uri;
     }