+++ t-relationalToRdf/src/main/java/eu/unifiedviews/plugins/transformer/relationaltordf/column/ValueGenerator.java
@@ -3,9 +3,9 @@ package eu.unifiedviews.plugins.transformer.relationaltordf.column;
 import java.util.List;
 import java.util.Map;
 
-import org.openrdf.model.URI;
-import org.openrdf.model.Value;
-import org.openrdf.model.ValueFactory;
+import org.eclipse.rdf4j.model.IRI;
+import org.eclipse.rdf4j.model.Value;
+import org.eclipse.rdf4j.model.ValueFactory;
 
 import eu.unifiedviews.plugins.transformer.relationaltordf.ConversionFailed;
 
@@ -34,8 +34,8 @@ public interface ValueGenerator {
     public Value generateValue(List<Object> row, ValueFactory valueFactory);
 
     /**
-     * @return URI for generated value.
+     * @return IRI for generated value.
      */
-    public URI getUri();
+    public IRI getUri();
 
 }