+++ e-httpRequest/CHANGELOG.md
@@ -1,8 +1,9 @@
 E-HttpRequest
 ----------
 
-v1.1.0
+v2.0.0
 ---
+* Updated to use UnifiedViews helpers 3.X with rdf4j support
 * Support for HTTPS
 * Support for dynamic configuration (only request for raw mode)
 * Content type may be customized in the file mode