+++ t-filterValidXml/CHANGELOG.md
@@ -1,6 +1,10 @@
 T-FilterValidXml
 ----------
 
+v3.0.0
+---
+* Updated to use UnifiedViews helpers 3.X with rdf4j support
+
 v2.1.3
 ---
 * Update to helpers 2.1.3