+++ t-sparqlConstruct/CHANGELOG.md
@@ -1,16 +1,20 @@
 T-SparqlConstruct
 ----------
 
-2.1.6
+v3.0.0
+---
+* Updated to use UnifiedViews helpers 3.X with rdf4j support
+
+v2.1.6
 ---
 * Added support for [RDF Validation extension](https://grips.semantic-web.at/display/UDDOC/RDF+Validation)
 * Update to helpers 2.1.7
 
-2.1.5
+v2.1.5
 ---
 * Improved slovak translation
 
-2.1.4
+v2.1.4
 ---
 * Documentation (About) improved
 * Update to helpers 2.1.4